requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain-groq==0.1.0
python-dotenv==1.0.0
playwright==1.42.0
pydantic>=2.0.0